backend_clean/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets>=12.0
wsproto>=1.2.0
python-multipart==0.0.17
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic[email]==2.9.2

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# ─── Authentication ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── Task Queue ───────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.2.1

# ─── Computer Vision ──────────────────────────────────────────────────────────
ultralytics==8.3.40           # YOLOv8 object detection
opencv-python-headless==4.10.0.84

# ─── Audio Processing ─────────────────────────────────────────────────────────
openai-whisper>=20240930                   # Fallback ASR (always installed for safety)
faster-whisper>=1.0.3                      # v5 primary ASR — 4-8× faster + Silero VAD
ctranslate2>=4.4.0                         # faster-whisper inference backend
librosa>=0.10.0
soundfile>=0.12.0
ffmpeg-python>=0.2.0
noisereduce>=3.0.3                         # Pre-Whisper denoiser

# ─── Deep Learning (CPU-only) ─────────────────────────────────────────────────
# Install CPU versions to avoid large GPU dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
# For GPU training on Colab/Kaggle, use:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# ─── HuggingFace / Transformers ───────────────────────────────────────────────
transformers>=4.20.0
accelerate>=0.20.0
huggingface-hub>=0.15.0

# ─── ONNX Runtime (CPU inference for quantized models) ────────────────────────
# onnxruntime>=1.14.0

# ─── TensorFlow (YAMNet audio classifier) ─────────────────────────────────────
# tensorflow-cpu>=2.13.0
# tensorflow-hub>=0.15.0

# ─── Report Generation ────────────────────────────────────────────────────────
weasyprint==62.3
pandas==2.2.3

# ─── Utilities ────────────────────────────────────────────────────────────────
structlog==24.4.0
httpx==0.27.2
python-dotenv==1.0.1
numpy==1.26.4
pillow==11.0.0
aiofiles==24.1.0
